+++ packages/coin-sui/CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Change Log
 
 All notable changes to this project will be documented in this file.
@@ -8,3 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 
 - **coin-sui:** change sui derived path ([904c66c](https://github.com/okx/js-wallet-sdk/pull/4/commits/904c66caaad9c679f0d7263957109f6743265a00))
+
+# [1.0.2](https://github.com/okx/js-wallet-sdk) (2024-3-5)
+
+### New Features
+
+- **coin-sui:** support sui new private key format :suiprivkey****